Code Refractor category index

+++ app/Http/Livewire/Admin/Subcategory/Index.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Livewire\Admin\Subcategory;
 
+use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\Log;
 use App\Models\SubCategory;
 use Livewire\Component;
@@ -13,14 +15,22 @@ class Index extends Component
     use WithFileUploads;
     use WithPagination;
 
+    public SubCategory $subcategory;
     public $img;
     public $search;
     public $readyToLoad = false;
-    public SubCategory $subcategory;
 
-    protected $paginationTheme = 'bootstrap';
     protected $queryString = ['search'];
+    /**
+     * @var string
+     * Site front type
+     */
+    protected $paginationTheme = 'bootstrap';
 
+    /**
+     * @var string[]
+     * Input rules
+     */
     protected $rules = [
         'subcategory.title' => 'required|min:3',
         'subcategory.name' => 'required',
@@ -36,6 +46,12 @@ class Index extends Component
     }
 
 
+    public function loadCategory()
+    {
+        $this->readyToLoad = true;
+    }
+
+
     /**
      * @throws \Illuminate\Validation\ValidationException
      */
@@ -45,35 +61,45 @@ class Index extends Component
     }
 
 
-    public function loadCategory()
-    {
-        $this->readyToLoad = true;
-    }
-
-
     public function categoryForm()
     {
         $this->validate();
+
+        $subCategory = SubCategory::query()->create([
+            'title' => $this->subcategory->title,
+            'name' => $this->subcategory->name,
+            'link' => $this->subcategory->link,
+            'parent' => $this->subcategory->parent,
+            'status' => $this->subcategory->status ? 1 : 0,
+        ]);
+
         if ($this->img) {
-            $this->subcategory->img = $this->uploadImage();
+            $subCategory->update([
+                'img' => $this->uploadImage()
+            ]);
         }
 
-        if (!$this->subcategory->status) {
-            $this->subcategory->status = 0;
-        }
-
-        $this->subcategory->save();
+        $this->subcategory->title = "";
+        $this->subcategory->name = "";
+        $this->subcategory->link = "";
+        $this->subcategory->parent = null;
+        $this->subcategory->status = false;
+        $this->img = null;
 
         Log::create([
             'user_id' => auth()->user()->id,
-            'url' => 'افزودن زیردسته ' . '-' . $this->subcategory->title,
+            'url' => 'افزودن زیر دسته' . '-' . $this->subcategory->title,
             'actionType' => 'ایجاد'
         ]);
-
         $this->emit('toast', 'success', ' زیردسته با موفقیت ایجاد شد.');
+
     }
 
 
+    /**
+     * @return string
+     * Upload image to memory
+     */
     public function uploadImage()
     {
         $year = now()->year;
@@ -91,13 +117,11 @@ class Index extends Component
         $category->update([
             'status' => 0
         ]);
-
         Log::create([
             'user_id' => auth()->user()->id,
-            'url' => 'غیرفعال کردن وضعیت  زیردسته' . '-' . $category->title,
+            'url' => 'غیرفعال کردن وضعیت زیر دسته' . '-' . $category->title,
             'actionType' => 'غیرفعال'
         ]);
-
         $this->emit('toast', 'success', 'وضعیت زیر دسته با موفقیت غیرفعال شد.');
     }
 
@@ -108,13 +132,11 @@ class Index extends Component
         $category->update([
             'status' => 1
         ]);
-
         Log::create([
             'user_id' => auth()->user()->id,
-            'url' => 'فعال کردن وضعیت زیردسته' . '-' . $category->title,
+            'url' => 'فعال کردن وضعیت زیر دسته' . '-' . $category->title,
             'actionType' => 'فعال'
         ]);
-
         $this->emit('toast', 'success', 'وضعیت زیر دسته با موفقیت فعال شد.');
     }
 
@@ -122,18 +144,26 @@ class Index extends Component
     public function deleteCategory($id)
     {
         $category = SubCategory::find($id);
-        $category->delete();
+        $childCategory = ChildCategory::where('parent', $id)->first();
+        if ($childCategory == null) {
+            $category->delete();
+            Log::create([
+                'user_id' => auth()->user()->id,
+                'url' => 'حذف کردن زیر دسته' . '-' . $category->title,
+                'actionType' => 'حذف'
+            ]);
+            $this->emit('toast', 'success', ' زیر دسته با موفقیت حذف شد.');
+        } else {
+            $this->emit('toast', 'success', ' امکان حذف وجود ندارد زیرا این دسته، شامل دسته کودک است!');
+        }
 
-        Log::create([
-            'user_id' => auth()->user()->id,
-            'url' => 'حذف کردن زیردسته ' . '-' . $category->title,
-            'actionType' => 'حذف'
-        ]);
-
-        $this->emit('toast', 'success', ' زیر دسته با موفقیت حذف شد.');
     }
 
 
+    /**
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * Send information to the site and load the site
+     */
     public function render()
     {
 
@@ -141,7 +171,6 @@ class Index extends Component
             ->orWhere('name', 'LIKE', "%{$this->search}%")
             ->orWhere('link', 'LIKE', "%{$this->search}%")
             ->latest()->paginate(10) : [];
-
         return view('livewire.admin.subcategory.index', compact('categories'));
     }
 }
